+++ ListableUI/Tests/Layout/Paged/PagedListLayoutTests.swift
@@ -104,6 +104,30 @@ class PagedListLayoutTests : XCTestCase
         
         return listView
     }
+    
+    func test_pagingStyle_defaultPeek() {
+        
+        let layout = PagedListLayout(
+            layoutAppearance: PagedAppearance(),
+            appearance: Appearance(),
+            behavior: Behavior(),
+            content: ListLayoutContent()
+        )
+        
+        XCTAssertEqual(layout.scrollViewProperties.pageScrollingBehavior, .full)
+    }
+    
+    func test_pagingStyle_customPeek() {
+        
+        let layout = PagedListLayout(
+            layoutAppearance: PagedAppearance(peek: PagedAppearance.Peek(value: 10)),
+            appearance: Appearance(),
+            behavior: Behavior(),
+            content: ListLayoutContent()
+        )
+        
+        XCTAssertEqual(layout.scrollViewProperties.pageScrollingBehavior, .peek)
+    }
 }
 
 
@@ -122,15 +146,77 @@ class PagedAppearanceTests : XCTestCase
 
 class PagedAppearance_PagingSize_Tests : XCTestCase
 {
-    func test_size()
+    func test_size_noPeek()
     {
         let size = CGSize(width: 30.0, height: 50.0)
         
-        XCTAssertEqual(PagedAppearance.PagingSize.view.size(for: size, direction: .vertical), CGSize(width: 30.0, height: 50.0))
-        XCTAssertEqual(PagedAppearance.PagingSize.view.size(for: size, direction: .horizontal), CGSize(width: 30.0, height: 50.0))
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(.none).size(for: size, isFirstItem: true, direction: .vertical),
+            CGSize(width: 30.0, height: 50.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(.none).size(for: size, isFirstItem: true, direction: .horizontal),
+            CGSize(width: 30.0, height: 50.0)
+        )
         
-        XCTAssertEqual(PagedAppearance.PagingSize.fixed(100).size(for: size, direction: .vertical), CGSize(width: 30.0, height: 100.0))
-        XCTAssertEqual(PagedAppearance.PagingSize.fixed(100).size(for: size, direction: .horizontal), CGSize(width: 100.0, height: 50.0))
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.fixed(100).size(for: size, isFirstItem: true, direction: .vertical),
+            CGSize(width: 30.0, height: 100.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.fixed(100).size(for: size, isFirstItem: true, direction: .horizontal),
+            CGSize(width: 100.0, height: 50.0)
+        )
+    }
+    
+    func test_size_uniformPeek() {
+        let size = CGSize(width: 30.0, height: 50.0)
+        let peek = PagedAppearance.Peek(
+            value: 10,
+            firstItemConfiguration: .uniform
+        )
+        
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: true, direction: .vertical),
+            CGSize(width: 30.0, height: 30.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: false, direction: .vertical),
+            CGSize(width: 30.0, height: 30.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: true, direction: .horizontal),
+            CGSize(width: 10.0, height: 50.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: false, direction: .horizontal),
+            CGSize(width: 10.0, height: 50.0)
+        )
+    }
+    
+    func test_size_customLeadingPeek() {
+        let size = CGSize(width: 30.0, height: 50.0)
+        let peek = PagedAppearance.Peek(
+            value: 10,
+            firstItemConfiguration: .customLeading(5)
+        )
+        
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: true, direction: .vertical),
+            CGSize(width: 30.0, height: 35.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: false, direction: .vertical),
+            CGSize(width: 30.0, height: 30.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: true, direction: .horizontal),
+            CGSize(width: 15.0, height: 50.0)
+        )
+        XCTAssertEqual(
+            PagedAppearance.PagingSize.inset(peek).size(for: size, isFirstItem: false, direction: .horizontal),
+            CGSize(width: 10.0, height: 50.0)
+        )
     }
 }